refactor to use details
